GCC-2.95 and x86-64 VXD compatibility changes.
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@1709 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/arch/linux/async/debug.c
@@ -67,14 +67,14 @@ static void stop_gdb(void)
 static void print_trace (void)
 {
   void *array[10];
-  size_t size;
+  int size;
   char **strings;
   size_t i;
   
   size = backtrace (array, 10);
   strings = backtrace_symbols (array, size);
   
-  printf ("Obtained %zd stack frames.\n", size);
+  printf ("Obtained %d stack frames.\n", size);
   
   for (i = 0; i < size; i++)
     printf ("%s\n", strings[i]);
